Fix bugs

+++ bot/handlers.py
@@ -129,7 +129,8 @@ async def handle_interrupting_start(message: Message, state: FSMContext) -> None
 @game_router.callback_query(Text('main'), GameStates.interruption)
 async def handle_interruption_confirmation(callback: CallbackQuery, state: FSMContext, bot: Bot) -> None:
     user_data = await state.get_data()
-    score1, score2, hits1, hits2 = user_data['score1'], user_data['score2'], user_data['hits1'], user_data['hits2']
+    score1, score2 = user_data.get('score1'), user_data.get('score2')
+    hits1, hits2 = user_data.get('hits1'), user_data.get('hits2')
     with suppress(TelegramBadRequest):
         if 'interrupting_message_id' in user_data:
             await bot.delete_message(chat_id=callback.message.chat.id, message_id=user_data['interrupting_message_id'])
@@ -137,17 +138,17 @@ async def handle_interruption_confirmation(callback: CallbackQuery, state: FSMCo
             await bot.delete_message(chat_id=callback.message.chat.id, message_id=user_data['new_message_id'])
     if 'pinned_message_id' in user_data:
         await bot.unpin_chat_message(chat_id=callback.message.chat.id, message_id=user_data['pinned_message_id'])
-    if 'backup_state' in user_data and user_data['backup_state'] == GameStates.answer:
+    if user_data.get('backup_state') == GameStates.answer:
         async with aiohttp.ClientSession() as session:
             async with session.post(
                 f'{API}/finish',
                 json={
                     'round_id': user_data['round_id'],
                     'abort': True,
-                    'score1': score1,
-                    'score2': score2,
-                    'hits1': hits1,
-                    'hits2': hits2
+                    'score1': int(score1),
+                    'score2': int(score2),
+                    'hits1': int(hits1),
+                    'hits2': int(hits2)
                 }
             ):
                 logger.info(f'GAME Aborted round {user_data["round_id"]} by {format_player(user_data["player"])}')
@@ -164,7 +165,8 @@ async def handle_interruption_cancel(callback: CallbackQuery, state: FSMContext)
     user_data = await state.get_data()
     await callback.message.delete()
     await callback.answer()
-    await user_data['interrupting_message'].delete()
+    if user_data.get('interrupting_message'):
+        await user_data['interrupting_message'].delete()
     await state.set_state(user_data['backup_state'])
